Add Context interfaces
It would be nice to instrument the start/stop process with tracing. To do that, we need to feed a `Context` object down from the start to all of the objects being started or stopped. Those objects can be responsible for adding spans if they desire.

This adds `StartContext` and `StopContext`, which work with `ContextStarter` and `ContextStopper` interfaces as well as the original `Starter` and `Stopper` interfaces. If the `Context*` interfaces don't exist, the original interfaces are called instead. Otherwise, the `Context*` interfaces are preferred.

Note I added tests and ran them locally, but nothing runs these in CI right now